csrf protection to journalindex delete related #609

+++ src/Ojs/JournalBundle/Controller/JournalIndexController.php
@@ -9,6 +9,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Ojs\Common\Controller\OjsController as Controller;
 use Ojs\JournalBundle\Entity\JournalIndex;
 use Ojs\JournalBundle\Form\JournalIndexType;
+use Symfony\Component\Security\Core\Exception\TokenNotFoundException;
 
 /**
  * JournalIndex controller.
@@ -27,6 +28,8 @@ class JournalIndexController extends Controller
         $grid = $this->get('grid')->setSource($source);
 
         $actionColumn = new ActionsColumn("actions", 'actions');
+        ActionHelper::setup($this->get('security.csrf.token_manager'));
+
         $rowAction[] = ActionHelper::showAction('admin_journalindex_show', 'id');
         $rowAction[] = ActionHelper::editAction('admin_journalindex_edit', 'id');
         $rowAction[] = ActionHelper::deleteAction('admin_journalindex_delete', 'id');
@@ -197,10 +200,15 @@ class JournalIndexController extends Controller
      * @param  JournalIndex                                       $entity
      * @return \Symfony\Component\HttpFoundation\RedirectResponse
      */
-    public function deleteAction(JournalIndex $entity)
+    public function deleteAction(Request $request, JournalIndex $entity)
     {
         $this->throw404IfNotFound($entity);
         $em = $this->getDoctrine()->getManager();
+
+        $csrf = $this->get('security.csrf.token_manager');
+        $token = $csrf->getToken('admin_journalindex'.$entity->getId());
+        if($token!=$request->get('_token'))
+            throw new TokenNotFoundException("Token Not Found!");
         $em->remove($entity);
         $em->flush();
         $this->successFlashBag('successful.remove');
